Refactor app, config, and database code.
app
===
Split AppRequest and ServerAddress out off app.go into appropriately
named files.

app/config
==========
Move config handling to a subpackage, app/config, and update remaining
code to reflect change.

app/database
============
Move definition of database and associated table management primitives
to a new app/database subpackage, and split the code for the different
primitives out into their own files.

Move the database table definitions into the app/database subpackage.

Add functions to retrieve references to the TableSpec definitions of
database tables in both Operational and Telemetry DBs.

Create a new abstraction, AppDb, to manage the database connection and
tables, and use it in place of the DbConnection in the App structure,
and when calling SetupDB for table rows.

Also pass a reference to the DBConfig, rather than the DBConfig itself,
to the DbConnection.Setup() call.

app/database/dbmanager
======================
Move app/dbmanager to app/database/dbmanager subpackage.

app/database/dialect
====================
Move SQL specific dialect placeholder management to a new subpackage,
app/database/dialect.

app/database/operationaldb
==========================
Move definition and setup of the Operational DB to a new subpackage,
app/database/operationaldb.

app/database/telemetrydb
========================
Move definition and setup of the Telemetry DB to a new subpackage,
app/database/telemetrydb.
