+++ app/app.go
@@ -1,182 +1,30 @@
 package app
 
 import (
-	"compress/gzip"
-	"compress/zlib"
 	"context"
-	"encoding/json"
 	"errors"
-	"fmt"
-	"io"
 	"log/slog"
 	"net/http"
 	"os"
 	"os/signal"
-	"strings"
 	"syscall"
 	"time"
 
+	"github.com/SUSE/telemetry-server/app/config"
+	"github.com/SUSE/telemetry-server/app/database"
+	"github.com/SUSE/telemetry-server/app/database/operationaldb"
+	"github.com/SUSE/telemetry-server/app/database/telemetrydb"
 	"github.com/SUSE/telemetry/pkg/logging"
 	_ "github.com/mattn/go-sqlite3"
 )
-
-// ServerAddress is a struct tracking the server address
-type ServerAddress struct {
-	Hostname string
-	Port     int
-}
-
-func (s ServerAddress) String() string {
-	return fmt.Sprintf("%s:%d", s.Hostname, s.Port)
-}
-
-func (s *ServerAddress) Setup(api APIConfig) {
-	s.Hostname, s.Port = api.Host, api.Port
-}
-
-// AppVars is a map var name to value
-type AppVars map[string]string
-
-// AppRequest is a struct tracking the resources associated with handling a request
-type AppRequest struct {
-	W     http.ResponseWriter
-	R     *http.Request
-	Vars  AppVars
-	Log   *slog.Logger
-	Quiet bool
-}
-
-func (ar *AppRequest) getReader() (io.ReadCloser, error) {
-	// Check the Content-Encoding header
-	switch ar.R.Header.Get("Content-Encoding") {
-	case "gzip":
-		return gzip.NewReader(ar.R.Body)
-	case "deflate":
-		return zlib.NewReader(ar.R.Body)
-	default:
-		return ar.R.Body, nil
-	}
-
-}
-
-func ReqLogger(r *http.Request) *slog.Logger {
-	return slog.Default().With(slog.String("method", r.Method), slog.Any("URL", r.URL))
-}
-
-func NewAppRequest(w http.ResponseWriter, r *http.Request, v AppVars) *AppRequest {
-	return &AppRequest{
-		W:    w,
-		R:    r,
-		Vars: v,
-		Log:  ReqLogger(r),
-	}
-}
-
-func QuietAppRequest(w http.ResponseWriter, r *http.Request, v AppVars) (ar *AppRequest) {
-	ar = NewAppRequest(w, r, v)
-	ar.Quiet = true
-	return
-}
-
-func (ar *AppRequest) GetHeader(header string) (value string) {
-	value = ar.R.Header.Get(header)
-	ar.Log.Debug("Request header", slog.String(header, value))
-	return
-}
-
-func (ar *AppRequest) GetAuthorization() string {
-	return ar.GetHeader("Authorization")
-}
-
-func (ar *AppRequest) GetAuthToken() string {
-	return strings.TrimPrefix(ar.GetAuthorization(), "Bearer ")
-}
-
-func (ar *AppRequest) GetRegistrationId() string {
-	return ar.GetHeader("X-Telemetry-Registration-Id")
-}
-
-func (ar *AppRequest) SetHeader(header, value string) {
-	ar.Log.Debug("Response header", slog.String(header, value))
-	ar.W.Header().Set(header, value)
-}
-
-func (ar *AppRequest) ContentType(contentType string) {
-	ar.SetHeader("Content-Type", contentType)
-}
-
-func (ar *AppRequest) ContentTypeJSON() {
-	ar.ContentType("application/json")
-}
-
-func (ar *AppRequest) SetWwwAuthenticate(challenge, realm, scope string) {
-	ar.SetHeader(
-		"WWW-Authenticate",
-		fmt.Sprintf(`%s realm="%s" scope="%s"`, challenge, realm, scope),
-	)
-}
-
-func (ar *AppRequest) SetWwwAuthScope(scope string) {
-	ar.SetWwwAuthenticate("Bearer", "suse-telemetry-service", scope)
-}
-
-func (ar *AppRequest) SetWwwAuthReauth() {
-	ar.SetWwwAuthScope("authenticate")
-}
-
-func (ar *AppRequest) SetWwwAuthRegister() {
-	ar.SetWwwAuthScope("register")
-}
-
-func (ar *AppRequest) Status(statusCode int) {
-	ar.Log.Debug("Response status", slog.Int("code", statusCode))
-	ar.W.WriteHeader(statusCode)
-}
-
-func (ar *AppRequest) StatusInternalServerError() {
-	ar.Status(http.StatusInternalServerError)
-}
-
-func (ar *AppRequest) Write(data []byte) (code int, err error) {
-	ar.Log.Debug("Response write", slog.Int("length", len(data)))
-	code, err = ar.W.Write(data)
-	return
-}
-
-func (ar *AppRequest) ErrorResponse(code int, errorMessage string) {
-	ar.Log.Debug("Setting error response", slog.Int("code", code), slog.String("error", errorMessage))
-	ar.JsonResponse(code, map[string]string{"error": errorMessage})
-}
-
-func (ar *AppRequest) JsonResponse(code int, payload any) {
-	respContent, err := json.Marshal(payload)
-	if err != nil {
-		ar.ErrorResponse(http.StatusInternalServerError, err.Error())
-		ar.Log.Error("Payload marshal failed", slog.Int("code", code), slog.String("error", err.Error()))
-		return
-	}
-
-	ar.ContentTypeJSON()
-	ar.Status(code)
-	writeCode, err := ar.Write(respContent)
-	if err != nil {
-		ar.Log.Error("Response write failed", slog.Int("code", code), slog.Int("writeCode", writeCode), slog.String("error", err.Error()))
-	} else {
-		if !ar.Quiet {
-			ar.Log.Info("Response", slog.Int("code", code))
-		} else {
-			ar.Log.Debug("Response", slog.Int("code", code))
-		}
-	}
-}
 
 // App is a struct tracking the resources associated with the application
 type App struct {
 	// public
 	Name          string
-	Config        *Config
-	TelemetryDB   DbConnection
-	OperationalDB DbConnection
+	Config        *config.Config
+	TelemetryDB   *database.AppDb
+	OperationalDB *database.AppDb
 	Address       ServerAddress
 	Handler       http.Handler
 	LogManager    *logging.LogManager
@@ -188,7 +36,7 @@ type App struct {
 	debugMode bool
 }
 
-func NewApp(name string, cfg *Config, handler http.Handler, debugMode bool) *App {
+func NewApp(name string, cfg *config.Config, handler http.Handler, debugMode bool) *App {
 	var err error
 
 	a := new(App)
@@ -204,12 +52,14 @@ func NewApp(name string, cfg *Config, handler http.Handler, debugMode bool) *App
 		panic(err)
 	}
 
-	// setup databases
-	err = a.TelemetryDB.Setup("Telemetry", cfg.DataBases.Telemetry)
+	// setup telemetry database
+	a.OperationalDB, err = operationaldb.New(cfg)
 	if err != nil {
 		panic(err)
 	}
-	err = a.OperationalDB.Setup("Operational", cfg.DataBases.Operational)
+
+	// setup telemetry database
+	a.TelemetryDB, err = telemetrydb.New(cfg)
 	if err != nil {
 		panic(err)
 	}
@@ -260,31 +110,21 @@ func (a *App) ListenOn() string {
 	return a.Address.String()
 }
 
-func (a *App) Initialize() error {
+func (a *App) Initialize() (err error) {
 
-	// operational DB setup
-	if err := a.OperationalDB.Connect(); err != nil {
+	// operational DB connect & setup
+	if err = a.OperationalDB.Connect(); err != nil {
 		slog.Error("Operational DB connection setup failed", slog.String("error", err.Error()))
-		return err
+		return
 	}
 
-	if err := a.OperationalDB.EnsureTableSpecsExist(operationalTables); err != nil {
-		slog.Error("Operational DB tables setup failed", slog.String("error", err.Error()))
-		return err
-	}
-
-	// telemetry DB setup
-	if err := a.TelemetryDB.Connect(); err != nil {
+	// telemetry DB connect & setup
+	if err = a.TelemetryDB.Connect(); err != nil {
 		slog.Error("Telemetry DB connection setup failed", slog.String("error", err.Error()))
-		return err
+		return
 	}
 
-	if err := a.TelemetryDB.EnsureTableSpecsExist(telemetryTables); err != nil {
-		slog.Error("Telemetry DB tables setup failed", slog.String("error", err.Error()))
-		return err
-	}
-
-	return nil
+	return
 }
 
 func (a *App) ListenAndServe() (err error) {
@@ -318,18 +158,30 @@ func (a *App) Shutdown() (err error) {
 	slog.Debug("Succeeded in shutdown of server")
 
 	// close the DB connections
-	dbConns := []DbConnection{
+	adbs := []*database.AppDb{
 		a.TelemetryDB,
 		a.OperationalDB,
 	}
-	for _, dbConn := range dbConns {
-		slog.Debug("Attempting " + dbConn.name + " DB close...")
-		if err = dbConn.dbMgr.Close(); err != nil {
-			slog.Debug(dbConn.name+" DB close failed", slog.String("err", err.Error()))
+	for _, adb := range adbs {
+		slog.Debug(
+			"Attempting DB close",
+			slog.String("database", adb.Name()),
+		)
+		if err = adb.Close(); err != nil {
+			slog.Error(
+				"DB close failed",
+				slog.String("database", adb.Name()),
+				slog.String("err", err.Error()),
+			)
 			return
 		}
-		slog.Debug("Succeeded in close of " + dbConn.name + " DB")
+		slog.Info(
+			"Closed DB",
+			slog.String("database", adb.Name()),
+		)
 	}
+
+	slog.Info("Shutdown complete")
 
 	return
 }
