+++ app/database/const.go
@@ -1,0 +1,5 @@
+package database
+
+const (
+	CREATE_TABLE_ADVISORY = 3141592653589793
+)
