+++ app/database/dialect/placeholder_test.go
@@ -1,4 +1,4 @@
-package app
+package dialect
 
 import (
 	"testing"
