+++ app/database/tagsets.go
@@ -1,0 +1,102 @@
+package database
+
+import (
+	"database/sql"
+	"encoding/json"
+	"log/slog"
+)
+
+var tagSetsTableSpec = TableSpec{
+	Name: "tagSets",
+	Columns: []TableSpecColumn{
+		{Name: "id", Type: "INTEGER", PrimaryKey: true, Identity: true},
+		{Name: "tagSet", Type: "VARCHAR"},
+	},
+}
+
+func GetTagSetsTableSpec() *TableSpec {
+	return &tagSetsTableSpec
+}
+
+type TagSetRow struct {
+	TableRowCommon
+
+	Id     int64  `json:"id"`
+	TagSet string `json:"tagSet"`
+}
+
+func (t *TagSetRow) Init(tagSet string) {
+	t.TagSet = tagSet
+}
+
+func (t *TagSetRow) String() string {
+	bytes, _ := json.Marshal(t)
+	return string(bytes)
+}
+
+func (t *TagSetRow) SetupDB(adb *AppDb) error {
+	t.tableSpec = GetTagSetsTableSpec()
+	return t.TableRowCommon.SetupDB(adb)
+}
+
+func (t *TagSetRow) Exists() bool {
+	stmt, err := t.SelectStmt(
+		[]string{
+			"id",
+		},
+		[]string{
+			"tagSet",
+		},
+		SelectOpts{}, // no special options
+	)
+	if err != nil {
+		slog.Error(
+			"exists statement generation failed",
+			slog.String("table", t.TableName()),
+			slog.String("error", err.Error()),
+		)
+		panic(err)
+	}
+
+	row := t.DB().QueryRow(stmt, t.TagSet)
+	if err := row.Scan(&t.Id); err != nil {
+		if err != sql.ErrNoRows {
+			slog.Error("tagSet existence check failed", slog.String("tagSet", t.TagSet), slog.String("error", err.Error()))
+		}
+		return false
+	}
+	return true
+}
+
+func (t *TagSetRow) Insert() (err error) {
+	stmt, err := t.InsertStmt(
+		[]string{
+			"tagSet",
+		},
+		"id",
+	)
+	if err != nil {
+		slog.Error(
+			"exists statement generation failed",
+			slog.String("table", t.TableName()),
+			slog.String("error", err.Error()),
+		)
+		return
+	}
+	row := t.DB().QueryRow(
+		stmt,
+		t.TagSet,
+	)
+	if err = row.Scan(
+		&t.Id,
+	); err != nil {
+		slog.Error(
+			"insert failed",
+			slog.String("table", t.TableName()),
+			slog.String("tagSet", t.TagSet),
+			slog.String("error", err.Error()),
+		)
+	}
+
+	return
+}
