+++ app/database/telemetrydb/database.go
@@ -1,0 +1,25 @@
+package telemetrydb
+
+import (
+	"github.com/SUSE/telemetry-server/app/config"
+	"github.com/SUSE/telemetry-server/app/database"
+)
+
+// Telemetry DB Tables
+var telemetryDbTables = database.DbTables{
+	database.GetCustomersTableSpec(),
+	database.GetTagSetsTableSpec(),
+	database.GetTelemetryTableSpec(),
+}
+
+func GetTables() database.DbTables {
+	return telemetryDbTables
+}
+
+func New(cfg *config.Config) (*database.AppDb, error) {
+	return database.GetDb(
+		"Telemetry",
+		&cfg.DataBases.Telemetry,
+		GetTables(),
+	)
+}
