+++ app/handler_authenticate.go
@@ -6,6 +6,7 @@ import (
 	"log/slog"
 	"net/http"
 
+	"github.com/SUSE/telemetry-server/app/database"
 	"github.com/SUSE/telemetry/pkg/restapi"
 )
 
@@ -36,8 +37,8 @@ func (a *App) AuthenticateClient(ar *AppRequest) {
 	ar.Log.Debug("Unmarshaled", slog.Any("caReq", &caReq))
 
 	// register the client
-	client := new(ClientsRow)
-	if err = client.SetupDB(&a.OperationalDB); err != nil {
+	client := new(database.ClientsRow)
+	if err = client.SetupDB(a.OperationalDB); err != nil {
 		ar.Log.Error("clientsRow.SetupDB() failed", slog.String("error", err.Error()))
 		ar.ErrorResponse(http.StatusInternalServerError, "failed to access DB")
 		return
