+++ app/handler_healthcheck.go
@@ -12,14 +12,14 @@ func (a *App) HealthCheck(ar *AppRequest) {
 
 func (a *App) LiveCheck(ar *AppRequest) {
 	ar.Log.Debug("Checking liveness probe")
-	err := a.TelemetryDB.dbMgr.Ping()
+	err := a.TelemetryDB.Ping()
 	if err != nil {
 		ar.Log.Error("Failed liveness probe")
 		ar.JsonResponse(http.StatusInternalServerError, `{"live": false}`)
 		return
 	}
 
-	err = a.OperationalDB.dbMgr.Ping()
+	err = a.OperationalDB.Ping()
 	if err != nil {
 		ar.Log.Error("Failed liveness probe")
 		ar.JsonResponse(http.StatusInternalServerError, `{"live": false}`)
