--- app/tables.go
@@ -1,14 +1,0 @@
-package app
-
-// Operational DB Tables
-var operationalTables = []TableSpec{
-	reportsStagingTableSpec,
-	clientsTableSpec,
-}
-
-// Telemetry DB Tables
-var telemetryTables = []TableSpec{
-	customersTableSpec,
-	tagSetsTableSpec,
-	telemetryTableSpec,
-}
