+++ app/tagsets.go
@@ -1,10 +1,7 @@
 package app
 
 import (
-	"database/sql"
-	"encoding/json"
 	"fmt"
-	"log/slog"
 	"slices"
 	"strings"
 )
@@ -39,95 +36,4 @@ func createTagSet(tags []string) string {
 	// append the bundle tags to the data item tags
 	uniqueTags := uniqueSortTags(tags)
 	return fmt.Sprintf("%s%s%s", tagSetSep, strings.Join(uniqueTags, tagSetSep), tagSetSep)
-}
-
-var tagSetsTableSpec = TableSpec{
-	Name: "tagSets",
-	Columns: []TableSpecColumn{
-		{Name: "id", Type: "INTEGER", PrimaryKey: true, Identity: true},
-		{Name: "tagSet", Type: "VARCHAR"},
-	},
-}
-
-type TagSetRow struct {
-	TableRowCommon
-
-	Id     int64  `json:"id"`
-	TagSet string `json:"tagSet"`
-}
-
-func (t *TagSetRow) Init(tagSet string) {
-	t.TagSet = tagSet
-}
-
-func (t *TagSetRow) String() string {
-	bytes, _ := json.Marshal(t)
-	return string(bytes)
-}
-
-func (t *TagSetRow) SetupDB(db *DbConnection) error {
-	t.tableSpec = &tagSetsTableSpec
-	return t.TableRowCommon.SetupDB(db)
-}
-
-func (t *TagSetRow) Exists() bool {
-	stmt, err := t.SelectStmt(
-		[]string{
-			"id",
-		},
-		[]string{
-			"tagSet",
-		},
-		SelectOpts{}, // no special options
-	)
-	if err != nil {
-		slog.Error(
-			"exists statement generation failed",
-			slog.String("table", t.TableName()),
-			slog.String("error", err.Error()),
-		)
-		panic(err)
-	}
-
-	row := t.DB().QueryRow(stmt, t.TagSet)
-	if err := row.Scan(&t.Id); err != nil {
-		if err != sql.ErrNoRows {
-			slog.Error("tagSet existence check failed", slog.String("tagSet", t.TagSet), slog.String("error", err.Error()))
-		}
-		return false
-	}
-	return true
-}
-
-func (t *TagSetRow) Insert() (err error) {
-	stmt, err := t.InsertStmt(
-		[]string{
-			"tagSet",
-		},
-		"id",
-	)
-	if err != nil {
-		slog.Error(
-			"exists statement generation failed",
-			slog.String("table", t.TableName()),
-			slog.String("error", err.Error()),
-		)
-		return
-	}
-	row := t.DB().QueryRow(
-		stmt,
-		t.TagSet,
-	)
-	if err = row.Scan(
-		&t.Id,
-	); err != nil {
-		slog.Error(
-			"insert failed",
-			slog.String("table", t.TableName()),
-			slog.String("tagSet", t.TagSet),
-			slog.String("error", err.Error()),
-		)
-	}
-
-	return
 }
