+++ server/telemetry-server/main.go
@@ -9,6 +9,7 @@ import (
 	"net/http"
 
 	"github.com/SUSE/telemetry-server/app"
+	"github.com/SUSE/telemetry-server/app/config"
 	"github.com/SUSE/telemetry/pkg/logging"
 	"github.com/gorilla/mux"
 )
@@ -38,9 +39,9 @@ func (rw *routerWrapper) healthCheck(w http.ResponseWriter, r *http.Request) {
 	rw.app.HealthCheck(app.QuietAppRequest(w, r, mux.Vars(r)))
 }
 
-func (rw *routerWrapper) liveCheck(w http.ResponseWriter, r *http.Request){
+func (rw *routerWrapper) liveCheck(w http.ResponseWriter, r *http.Request) {
 	rw.app.LiveCheck(app.QuietAppRequest(w, r, mux.Vars(r)))
-}	
+}
 
 // options is a struct of the options
 type options struct {
@@ -66,7 +67,7 @@ func main() {
 
 	slog.Debug("Preparing to start gorilla/mux based server", slog.Any("options", opts))
 
-	cfg := app.NewConfig(opts.Config)
+	cfg := config.NewConfig(opts.Config)
 	if err := cfg.Load(); err != nil {
 		slog.Error("config load failed", slog.String("config", opts.Config), slog.String("error", err.Error()))
 		panic(err)
@@ -81,7 +82,7 @@ func main() {
 
 func parseCommandLineFlags() {
 	// define available flags
-	flag.StringVar(&opts.Config, "config", app.DEFAULT_CONFIG, "Path to `config` file to use")
+	flag.StringVar(&opts.Config, "config", config.DEFAULT_CONFIG, "Path to `config` file to use")
 	flag.BoolVar(&opts.Debug, "debug", false, "Enables debug level messages")
 
 	// parse supplied command line flags
@@ -98,7 +99,7 @@ func SetupRouterWrapper(router *mux.Router, app *app.App) {
 	router.HandleFunc("/live", wrapper.liveCheck).Methods("GET", "HEAD")
 }
 
-func InitializeApp(cfg *app.Config, debug bool) (a *app.App, router *mux.Router) {
+func InitializeApp(cfg *config.Config, debug bool) (a *app.App, router *mux.Router) {
 	router = mux.NewRouter()
 
 	a = app.NewApp("Server", cfg, router, debug)
